+++ packages/blade/src/utils/makeMotionTime/makeMotionTime.native.ts
@@ -1,0 +1,5 @@
+const makeMotionTime = <T extends number>(time: T): T => {
+  return time;
+};
+
+export default makeMotionTime;